[feat]: liste des inventaires et details

+++ database/final/final.sql
@@ -175,8 +175,8 @@ CREATE TABLE inventaire_detail (
     id SERIAL PRIMARY KEY,
     id_inventaire INTEGER NOT NULL,
     id_materiel INTEGER NOT NULL,
-    quantite_initiale INTEGER NOT NULL,
-    quantite_finale INTEGER NOT NULL,
+    quantite_initiale DOUBLE PRECISION NOT NULL,
+    quantite_finale DOUBLE PRECISION NOT NULL,
     observations TEXT,
     CONSTRAINT fk_inventaire_detail_inventaire FOREIGN KEY (id_inventaire) REFERENCES inventaire(id) ON DELETE CASCADE
 );
